+++ application/src/main/java/de/dhbw/softwareengineering/PersonService.java
@@ -2,52 +2,71 @@ package de.dhbw.softwareengineering;
 
 import de.dhbw.softwareengineering.entities.Person;
 import de.dhbw.softwareengineering.exceptions.PersonNotFoundException;
+import de.dhbw.softwareengineering.mappers.PersonMapper;
 import de.dhbw.softwareengineering.repositories.PersonRepository;
-import de.dhbw.softwareengineering.values.Address;
-import de.dhbw.softwareengineering.values.Gender;
-import de.dhbw.softwareengineering.values.Name;
+import de.dhbw.softwareengineering.representations.PersonRepresentation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
 public class PersonService {
     private final PersonRepository personRepository;
+    private final PersonMapper personMapper;
+    private static final String EXCEPTION_MESSAGE = "No person with such a UUID.";
 
     @Autowired
-    public PersonService(PersonRepository personRepository) {
+    public PersonService(PersonRepository personRepository, PersonMapper personMapper) {
         this.personRepository = personRepository;
+        this.personMapper = personMapper;
     }
 
-    public Person create(Name name, Address address, LocalDate date, Gender gender){
-        Person person = new Person(name, address, date, gender);
+    public PersonRepresentation create(PersonRepresentation personRepresentation){
+        Person person = new Person(
+                personRepresentation.getId(),
+                personRepresentation.getName(),
+                personRepresentation.getAddress(),
+                personRepresentation.getBirthDate(),
+                personRepresentation.getGender()
+        );
 
-        return personRepository.save(person);
+        return personMapper.toPersonRepresentation(personRepository.insert(person));
     }
-    public Person get(UUID uuid) throws PersonNotFoundException{
-        return personRepository.findByUuid(uuid)
-                .orElseThrow(() -> new PersonNotFoundException("No person with such a UUID."));
-    }
-
-    public List<Person> getAll(){
-        return personRepository.findAll();
+    public PersonRepresentation get(UUID uuid) throws PersonNotFoundException{
+        return personMapper.toPersonRepresentation(personRepository.findByUuid(uuid)
+                .orElseThrow(() -> new PersonNotFoundException(EXCEPTION_MESSAGE)));
     }
 
-    public void delete(UUID uuid) throws PersonNotFoundException {
+    public List<PersonRepresentation> getAll(){
+
+        return personMapper.toPersonRepresentationList(personRepository.findAll());
+    }
+
+    public void delete(UUID uuid){
+        Optional<Person> optionalPerson = personRepository.findByUuid(uuid);
+
+        if(optionalPerson.isEmpty()){
+            throw new PersonNotFoundException(EXCEPTION_MESSAGE);
+        }
         personRepository.delete(uuid);
     }
 
-    public Person update(UUID uuid, Person updatedPerson){
-        Person person = get(uuid);
+    public PersonRepresentation update(UUID personUuid, PersonRepresentation updatedPerson){
+        Person person = getPerson(personUuid);
 
         person.setName(updatedPerson.getName());
-        person.setGender(updatedPerson.getGender());
         person.setAddress(updatedPerson.getAddress());
         person.setBirthDate(updatedPerson.getBirthDate());
+        person.setGender(updatedPerson.getGender());
 
-        return personRepository.save(person);
+        return personMapper.toPersonRepresentation(personRepository.save(person));
+    }
+
+    private Person getPerson(UUID uuid){
+        return personRepository.findByUuid(uuid)
+                .orElseThrow(() -> new PersonNotFoundException(EXCEPTION_MESSAGE));
     }
 }
